Add files via upload
What's the problem?
Sometimes (and only sometimes) when a customer places an order of 1 ticket, there will be 2 tickets stored in my database.

So, where is the problem?

Customers can make an order, Stripe will verify payment and then will send a webhook "charge.succeeded" to /stripe/webhook; this will trigger the function "handle" in OrderController.

After that, it will trigger the functions success in OrderController and create in TicketController. The problem is probably in create in TicketController.